Update symlinks root

+++ scripts/config_folder.sh
@@ -1,7 +1,7 @@
 source './osx/utils.sh'
 
 config_folder() {
-  # Get the directory where the script is located
+  # Get the directory where the script is located (SCRIPT_DIR is set by utils.sh)
   ROOT_DIR="$(git -C "${SCRIPT_DIR}" rev-parse --show-toplevel)"
   print_info "Symlinking configuration files to $HOME/.config directory."
 
@@ -11,10 +11,16 @@ config_folder() {
   fi
   
   for folder in "$ROOT_DIR/.config/"*; do
+    # Skip if glob didn't match anything
+    [ ! -e "$folder" ] && continue
+    
     folder_name=$(basename "$folder")
+    # Get absolute path (macOS compatible)
+    abs_folder_path="$(cd "$folder" && pwd)"
+    
     # Remove existing symlink if it exists
     [ -L "$HOME/.config/$folder_name" ] && rm "$HOME/.config/$folder_name"
-    execute "ln -s $(realpath "$folder") $HOME/.config/$folder_name"
+    execute "ln -s \"$abs_folder_path\" \"$HOME/.config/$folder_name\""
 
     print_info "Symlinked $folder_name to $HOME/.config/$folder_name"
   done
